requirements.txt
torch>=1.9.0
torchvision>=0.10.0
matplotlib>=3.4.3
Pillow>=8.4.0
numpy>=1.21.2
tqdm>=4.62.3
scikit-learn>=0.24.2
scipy>=1.7.1
